Add tile_single_column_menus option (KurzedMetal)
Some users complained that reading some menus was now more difficult,
since items in individual categories are displayed in two columns.
One example is when scanning a particular category of items in the
inventory; the player will need to scan left-right as well as up-down.

This new option allows users to force menus to use a single column. It
isn't a perfect solution, but webtiles and console already only support
single column menus, and there aren't too many complaints about that...